feat: 認証機能のファイルをfeaturesディレクトリに集約
